Spot patterns in strings

Spot patterns in strings.

string_spot(s, pattern)
s_spot(s, pattern)
spot(s, pattern)
string_spoti(s, pattern)
s_spoti(s, pattern)
spoti(s, pattern)
string_spotl(s, pattern)
s_spotl(s, pattern)
spotl(s, pattern)
string_spotm(s, pattern)
s_spotm(s, pattern)
spotm(s, pattern)
string_replace(s, search, replace)
s_replace(s, search, replace)
spot_replace(s, search, replace)
fr(s, search, replace)
search_replace(s, search, replace)
sr(s, search, replace)
string_remove(s, remove)
s_remove(s, remove)
string_countm(s, pattern, ...)
s_countm(s, pattern, ...)
countm(s, pattern, ...)

Arguments

s
A string (character) vector.

pattern
A regular expression pattern.

search
A pattern to search.

replace
The string to replace the searched pattern.

remove
The string to remove.

...
Parameters to pass to gregexpr

Value

Character vector for all functions except the spotl() and countm() functions, which produce Boolean and numeric vectors respectively.

Details

The function string_spot subsets a vector to the values matching a given pattern. Its synonyms are s_spot and spot.

The function string_spoti provides the indices where a pattern in a vector is found. Its synonyms are s_spoti and spoti.

The function string_spotl detects whether a pattern exists in a vector. Its synonyms are s_spotl and spotl.

The spotm functions spots pattern matches and returns NA if none are found.

The *_replace functions act the same as gsub.

The *_remove functions blank out a matching pattern.

The string_countm counts the number of matches in a string.

See also

Examples

rn <- rownames(mtcars)

string_spot(rn, "^M")

#>  [1] "Mazda RX4"     "Mazda RX4 Wag" "Merc 240D"     "Merc 230"     
#>  [5] "Merc 280"      "Merc 280C"     "Merc 450SE"    "Merc 450SL"   
#>  [9] "Merc 450SLC"   "Maserati Bora"

string_spoti(rn, "^M")

#>  [1]  1  2  8  9 10 11 12 13 14 31

string_spotl(rn, "^M")

#>  [1]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [13]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [25] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE

string_spotm(rn, "^M")

#>  [1] "Mazda RX4"     "Mazda RX4 Wag" NA              NA             
#>  [5] NA              NA              NA              "Merc 240D"    
#>  [9] "Merc 230"      "Merc 280"      "Merc 280C"     "Merc 450SE"   
#> [13] "Merc 450SL"    "Merc 450SLC"   NA              NA             
#> [17] NA              NA              NA              NA             
#> [21] NA              NA              NA              NA             
#> [25] NA              NA              NA              NA             
#> [29] NA              NA              "Maserati Bora" NA             

string_replace(rn, "^M", "Z")

#>  [1] "Zazda RX4"           "Zazda RX4 Wag"       "Datsun 710"         
#>  [4] "Hornet 4 Drive"      "Hornet Sportabout"   "Valiant"            
#>  [7] "Duster 360"          "Zerc 240D"           "Zerc 230"           
#> [10] "Zerc 280"            "Zerc 280C"           "Zerc 450SE"         
#> [13] "Zerc 450SL"          "Zerc 450SLC"         "Cadillac Fleetwood" 
#> [16] "Lincoln Continental" "Chrysler Imperial"   "Fiat 128"           
#> [19] "Honda Civic"         "Toyota Corolla"      "Toyota Corona"      
#> [22] "Dodge Challenger"    "AMC Javelin"         "Camaro Z28"         
#> [25] "Pontiac Firebird"    "Fiat X1-9"           "Porsche 914-2"      
#> [28] "Lotus Europa"        "Ford Pantera L"      "Ferrari Dino"       
#> [31] "Zaserati Bora"       "Volvo 142E"         

string_remove(rn, "^M")

#>  [1] "azda RX4"            "azda RX4 Wag"        "Datsun 710"         
#>  [4] "Hornet 4 Drive"      "Hornet Sportabout"   "Valiant"            
#>  [7] "Duster 360"          "erc 240D"            "erc 230"            
#> [10] "erc 280"             "erc 280C"            "erc 450SE"          
#> [13] "erc 450SL"           "erc 450SLC"          "Cadillac Fleetwood" 
#> [16] "Lincoln Continental" "Chrysler Imperial"   "Fiat 128"           
#> [19] "Honda Civic"         "Toyota Corolla"      "Toyota Corona"      
#> [22] "Dodge Challenger"    "AMC Javelin"         "Camaro Z28"         
#> [25] "Pontiac Firebird"    "Fiat X1-9"           "Porsche 914-2"      
#> [28] "Lotus Europa"        "Ford Pantera L"      "Ferrari Dino"       
#> [31] "aserati Bora"        "Volvo 142E"         

string_countm(rn, 'a')

#>  [1] 2 3 1 0 1 2 0 0 0 0 0 0 0 0 2 1 1 1 1 2 2 1 1 2 1 1 0 1 2 1 3 0